services/user/app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.5
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.9.2
python-dotenv==1.0.1
SQLAlchemy==2.0.35